Add files via upload
These files contain the source code for the C file for the scorecard. As this was my project for the first semester, many of the concepts that I usedd here may be amateur. That is why I made a version enhanced with AI(which is termed in the filename itself). The pdf file is a report on the same.
PS this is the first repository that I created on GitHub, so any suggestions are welcome!